pyproject.toml
[project]
name = "mailpilot"
version = "0.1.0"
description = "CRM application for cold email outreach via Gmail"
readme = "README.md"
authors = [
    { name = "Konstantin Borovik", email = "github@lab5.ca" }
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.0,<9.0.0",
    "google-api-python-client>=2.170.0,<3.0.0",
    "google-auth>=2.40.0,<3.0.0",
    "google-cloud-pubsub>=2.29.0,<3.0.0",
    "logfire>=4.0.0,<5.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "pydantic-ai-slim[anthropic]>=1.10.0,<2.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
]

[project.scripts]
mailpilot = "mailpilot.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "basedpyright>=1.37.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.13.0",
    "uv>=0.8.17",
]

[tool.basedpyright]
typeCheckingMode = "strict"
pythonVersion = "3.14"
include = ["src/mailpilot/", "tests/"]
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownLambdaType = false
reportUnusedFunction = "warning"

[tool.ruff]
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "C90",   # mccabe
    "COM",   # flake8-commas
    "D",     # pydocstyle
    "E",     # pycodestyle
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "PL",    # Pylint
    "PT",    # flake8-pytest-style
    "RUF",   # Ruff-specific rules
    "SIM",   # flake8-simplify
    "TID",   # flake8-tidy-imports
    "UP",    # pyupgrade
    "W",     # pycodestyle warnings
]
ignore = [
    "PLR2004",  # magic values in comparisons
    "PLC0415",  # import not at top-level (lazy imports)
    "D100",     # Missing docstring in public module
    "D104",     # Missing docstring in public package
    "D105",     # Missing docstring in magic method
    "D107",     # Missing docstring in __init__
    "COM812",   # trailing commas
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["mailpilot"]

[tool.ruff.lint.per-file-ignores]
"src/mailpilot/cli.py" = [
    "PLR0913", # Click commands have many parameters by design
]
"tests/**/*.py" = [
    "D",       # No docstrings required in tests
    "PLR2004", # Magic values OK in tests
    "PLR0913", # Too many arguments OK in tests
    "E501",    # Line too long
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::logfire._internal.config.LogfireNotConfiguredWarning",
]
